Move ReaderData to common
