+++ packages/readers_extra/src/weathergen/readers_extra/data_reader_cams.py
@@ -9,11 +9,10 @@ import xarray as xr
 from weathergen.datasets.data_reader_anemoi import _clip_lat, _clip_lon
 from weathergen.datasets.data_reader_base import (
     DataReaderTimestep,
-    ReaderData,
     TimeWindowHandler,
     TIndex,
-    check_reader_data,
 )
+from weathergen.common.data import ReaderData, check_reader_data
 
 ############################################################################
