+++ packages/readers_extra/src/weathergen/readers_extra/data_reader_eobs.py
@@ -17,12 +17,11 @@ from numpy.typing import NDArray
 
 from weathergen.datasets.data_reader_base import (
     DataReaderTimestep,
-    ReaderData,
     TimeWindowHandler,
     TIndex,
-    check_reader_data,
     str_to_timedelta,
 )
+from weathergen.common.data import ReaderData, check_reader_data
 
 _logger = logging.getLogger(__name__)
