+++ src/weathergen/datasets/data_reader_anemoi.py
@@ -19,11 +19,10 @@ from numpy.typing import NDArray
 
 from weathergen.datasets.data_reader_base import (
     DataReaderTimestep,
-    ReaderData,
     TimeWindowHandler,
     TIndex,
-    check_reader_data,
 )
+from weathergen.common.data import ReaderData, check_reader_data
 
 _logger = logging.getLogger(__name__)
